Поправил до конца

+++ Lab1_arch/Rectangle.cs
@@ -11,7 +11,7 @@ namespace Lab1_arch
 
         public override double Calculate(int n, int a, int b)
         {
-            double h = (b - a) / n;
+            double h = (double)((b - a)) / n;
             double res = 0;
             for (int i = 0; i < n; i++)
                 res += Function(a + h * (i + 0.5));
